`refactor(category): restructure package hierarchy and refactor DAO implementation`
