+++ src/main/java/br/com/libraryjdbc/model/dao/CategoryDAO.java
@@ -1,0 +1,14 @@
+package br.com.libraryjdbc.model.dao;
+
+import java.util.List;
+import br.com.libraryjdbc.model.entities.Category;
+
+public interface CategoryDAO {
+
+    Category save(Category category);
+    void update(Category category);
+    void remove(Long id);
+    Category findById(Long id);
+    List<Category> findAll();
+    Category findCategoryWithMostBooks();
+}
